Fixed several code hints and warnings
git-svn-id: svn://svn.code.sf.net/p/mobac/code/trunk/MOBAC@2797 a036c9c4-4c23-4c51-82c3-8fdb5c987102

+++ mobac/src/main/java/mobac/gui/components/JIntCombo.java
@@ -91,7 +91,7 @@ public class JIntCombo extends JComboBox<Integer> {
 			if (entry == null) {
 				return;
 			}
-			editorComponent.setValue(((Integer) entry).intValue(), true);
+			editorComponent.setValue((Integer) entry, true);
 		}
 
 		public void removeActionListener(ActionListener l) {
